Refactor database and table classes
Up to now there was a circular relationship between the database object and the
table objects. Now each table gets a StorageProxy which forwards read and write
operations to the storage and allows access only to the corresponding table data.

Also, the database object is now responsible for cleaning up the storage. The
table objects can no longer be used as a context manager (as suggested by
@krisztianfekete in #70). In addition, the database object will close the storage
if it is garbage collected and `close` hasn't been called yet.

See discussion in #70