Fix the Query constructor to not take parameters
In the old design (pre 3.0.0) a query was constructed using
`where('field')`. The new design instead uses `Query().field`. Due to
some confusion on my side the `Query` API docs still showed a usage
like `Query('field')` - which doesn't work anymore (see #117).
While the `Query` constructor still took a parameter, it was for
internal usage only. Now the parameter is removed to make it more clear
how `Query` is intended to be used.

While technically this is a breaking change, any code relying on the
`Query('field')` syntax didn't work anyway so this should not break
any existing code.

Closes #117