refactor(table): make `contains` accept a single doc ID only
When passing a list of document IDs to `Table.contains`, it's not clear what
this should mean – does it return `True` if *one* document ID exists? If *all*
document IDs are present? For this reason, this feature is removed here to
simplify the API and make it clear from the method interface what it does:
return `True` if a single document ID is present.