fix: make query callables work again
Passing a callable as a query used to work but got broken in v4.6.0
due to the necessity of having to distinguish between cacheable
and non-cacheable queries. This was implemented using a new
method named `is_cacheable` that was expected to be present on every
query object -- which of course is not the case for custom queries
that are implemented by passing a callable as a query.

We now only try to call `is_cacheable` if this method is present
on a query object. If it is NOT present, we treat it as cacheable.
This is to keep consistency with TinyDB's earlier behavior which
assumed that all queries are cacheable.

Fixes #454